Added unit test for WIC image loader
